BackEnd/requirements.txt
langchain
openai
faiss-cpu
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai

tiktoken

